Rename scheduling component to 'task'
