+++ src/task/cooperative_scheduler.rs
@@ -2,8 +2,7 @@ use alloc::{String, Vec, VecDeque};
 use core::mem;
 use core::ops::DerefMut;
 use core::sync::atomic::{AtomicUsize, Ordering};
-use scheduling::process;
-use scheduling::{DoesScheduling, Process, ProcessId, ProcessList, State, INIT_STK_SIZE};
+use task::{DoesScheduling, Process, ProcessId, ProcessList, State, INIT_STK_SIZE, process};
 use spin::RwLock;
 use syscall::error::Error;
