+++ src/task/process_list.rs
@@ -3,7 +3,7 @@ use alloc::btree_map::{self, BTreeMap};
 use alloc::Vec;
 use core::fmt;
 use core::result::Result;
-use scheduling::{Process, ProcessId, State};
+use task::{Process, ProcessId, State};
 use spin::RwLock;
 use syscall::error::Error;
